feat(starkex_handler): Moved signing method to utils
