+++ pragma-node/src/utils/mod.rs
@@ -2,7 +2,7 @@ pub use aws::build_pragma_signer_from_aws;
 pub use conversion::{convert_via_quote, format_bigdecimal_price, normalize_to_decimals};
 pub use custom_extractors::json_extractor::JsonExtractor;
 pub use custom_extractors::path_extractor::PathExtractor;
-pub use signing::starkex::{get_encoded_pair_id, get_entry_hash, get_external_asset_id};
+pub use signing::starkex::{get_encoded_pair_id, get_external_asset_id, sign_median_price};
 pub use signing::typed_data::TypedData;
 pub use types::UnixTimestamp;
